@@ -10,60 +10,59 @@
   ],
   "scripts": {
     "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "prepublish": "yarn run build",
+    "build": "babel src/ -d lib/ & react-app-rewired build",
+    "prepare": "npm run build",
     "test": "react-app-rewired test",
     "lint-staged": "lint-staged",
     "prettier": "npx prettier --write \"src/**/*.{js,jsx,ts,tsx,scss,less}\"",
     "eslint": "npx eslint --ext .js,jsx src",
     "stylelint": "stylelint \"src/**/*.{scss,less}\"",
-    "flow": "$(npm bin)/flow"
+    "flow": "flow"
   },
   "dependencies": {
-    "@commitlint/cli": "^8.3.4",
-    "@commitlint/config-conventional": "^8.3.4",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
     "@rowno/sparkline": "^4.0.0",
-    "antd": "^3.26.6",
-    "apollo-boost": "^0.4.7",
-    "apollo-cache-inmemory": "^1.6.5",
-    "apollo-client": "^2.6.8",
-    "apollo-link": "^1.2.13",
-    "apollo-link-context": "^1.0.19",
-    "apollo-link-error": "^1.1.12",
-    "apollo-link-http": "^1.5.16",
-    "auth0-js": "^9.12.1",
-    "bootstrap-css-only": "^4.3.1",
+    "antd": "^3.26.20",
+    "apollo-boost": "^0.4.9",
+    "apollo-cache-inmemory": "^1.6.6",
+    "apollo-client": "^2.6.10",
+    "apollo-link": "^1.2.14",
+    "apollo-link-context": "^1.0.20",
+    "apollo-link-error": "^1.1.13",
+    "apollo-link-http": "^1.5.17",
+    "auth0-js": "^9.14.0",
+    "bootstrap-css-only": "^4.4.1",
     "classnames": "^2.2.6",
-    "commitizen": "^4.0.3",
-    "commitlint": "^8.3.4",
-    "connected-react-router": "^6.6.1",
-    "cz-conventional-changelog": "^3.0.2",
+    "commitizen": "^3.1.1",
+    "commitlint": "^11.0.0",
+    "connected-react-router": "^6.8.0",
+    "cz-conventional-changelog": "^3.3.0",
     "enquire-js": "^0.2.1",
-    "eslint": "^6.8.0",
-    "flow-bin": "^0.104.0",
-    "graphql": "^14.5.8",
-    "graphql-tag": "^2.10.1",
+    "flow-bin": "^0.137.0",
+    "graphql": "^15.4.0",
+    "graphql-tag": "^2.11.0",
     "history": "^4.10.1",
-    "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.11",
-    "luxon": "^1.21.3",
+    "jwt-decode": "^3.0.0",
+    "lodash": "^4.17.20",
+    "luxon": "^1.25.0",
     "nprogress": "^0.2.0",
     "peity-react": "^0.0.7",
-    "qs": "^6.9.1",
-    "rc-drawer": "^2.1.1",
-    "react": "^16.12.0",
-    "react-apollo": "^2.5.8",
+    "qs": "^6.9.4",
+    "rc-drawer": "^4.1.0",
+    "react": "^16.13.1",
+    "react-apollo": "^3.1.5",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.13.1",
     "react-editmask": "^4.0.3",
-    "react-helmet": "^5.2.1",
-    "react-intl": "^2.9.0",
+    "react-helmet": "^6.1.0",
+    "react-intl": "^5.8.8",
     "react-loadable": "^5.5.0",
-    "react-redux": "^7.1.3",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
+    "react-redux": "^7.2.2",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-router-redux": "^4.0.8",
-    "react-scripts": "^3.3.0",
+    "react-scripts": "^3.4.3",
     "react-text-mask": "^5.4.3",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
