@@ -1,8 +1,9 @@
 import React, { Component } from 'react'
+import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
 import { Form, Row, Col, Radio, Input } from 'antd'
 
-import passengerActions from 'redux/passenger/actions'
+import passengerActions from 'redux/passengerList/actions'
 
 @Form.create()
 class PassengerFilter extends Component {
@@ -19,11 +20,19 @@ class PassengerFilter extends Component {
     this.handleChangeFilter = this.handleChangeFilter.bind(this)
   }
 
-  componentWillMount() {
-    const { getPassengers, filter, id } = this.props
+  componentDidMount() {
+    const {
+      getPassengers,
+      filter,
+      setFilter,
+      match: { params },
+    } = this.props
+    const { excursionId } = params
+
     const { status } = this.state
-    const payload = { ...filter, status, excursionId: id }
+    const payload = { ...filter, status, excursionId }
     getPassengers(payload)
+    setFilter(payload)
   }
 
   handleChangeStartPay(e) {
@@ -56,7 +65,7 @@ class PassengerFilter extends Component {
     const { id, passengers } = this.props
     const { status } = this.state
 
-    const excursion = passengers.find(x => x.id === id)
+    const excursion = passengers.find((x) => x.id === id)
 
     return (
       <Form layout="inline" className="form-filter">
@@ -122,14 +131,14 @@ class PassengerFilter extends Component {
   }
 }
 
-const mapStateToProps = ({ passenger: { filter, payload: passengers } }) => ({
+const mapStateToProps = ({ passengerList: { filter, payload: passengers } }) => ({
   filter,
   passengers,
 })
 
-const mapDispatchToProps = dispatch => ({
-  setFilter: filter => dispatch({ type: passengerActions.SET_STATE, filter }),
-  getPassengers: filter => dispatch({ type: passengerActions.GET_PASSENGERS, filter }),
+const mapDispatchToProps = (dispatch) => ({
+  setFilter: (filter) => dispatch({ type: passengerActions.SET_STATE, filter }),
+  getPassengers: (filter) => dispatch({ type: passengerActions.GET_PASSENGERS, filter }),
 })
 
-export default connect(mapStateToProps, mapDispatchToProps)(PassengerFilter)
+export default connect(mapStateToProps, mapDispatchToProps)(withRouter(PassengerFilter))
