@@ -1,29 +1,45 @@
-import React from 'react'
+import React, { useEffect } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
 import { Select, Divider } from 'antd'
+import actions from 'redux/customerList/actions'
 
-const CustomerSelect = ({ customers }) => (
-  <Select
-    size="default"
-    // TODO: max options show
-    showSearch
-    filterOption={(q, option) =>
-      q.split(' ').every(x => option.props.children.toLowerCase().includes(x))
-    }
-    dropdownRender={menu => (
-      <div>
-        {menu}
-        <Divider style={{ margin: '4px 0' }} />
-        <div style={{ padding: '8px', cursor: 'pointer' }}>
-          <i className="fa fa-plus" /> Novo cliente
-        </div>
-      </div>
-    )}
-  >
-    {customers.map(x => (
-      <Select.Option key={x.id} value={x.id}>
-        {`${x.name} - ${x.rg} - ${x.city}`}
-      </Select.Option>
-    ))}
-  </Select>
-)
-export default CustomerSelect
+export const CustomerSelect = ({ onChange }, ref) => {
+  const dispatch = useDispatch()
+
+  const { payload: customers } = useSelector(state => state.customerList)
+
+  useEffect(() => {
+    if (!customers || !customers.length)
+      dispatch({ type: actions.GET_CUSTOMERS })
+  }, [customers, dispatch])
+
+  return (
+    <Select
+      ref={ref}
+      size="default"
+      // TODO: max options show
+      showSearch
+      filterOption={(q, option) =>
+        q.split(' ').every(x => option.props.children.toLowerCase().includes(x))
+      }
+      onChange={onChange}
+      dropdownRender={menu => (
+        <>
+          {menu}
+          <Divider style={{ margin: '4px 0' }} />
+          <div style={{ padding: '8px', cursor: 'pointer' }}>
+            <i className="fa fa-plus" /> Novo cliente
+          </div>
+        </>
+      )}
+    >
+      {customers.map(x => (
+        <Select.Option key={x.id} value={x.id}>
+          {`${x.name} - ${x.document} - ${x.address.city}`}
+        </Select.Option>
+      ))}
+    </Select>
+  )
+}
+
+export default React.forwardRef(CustomerSelect)
