@@ -1,25 +1,15 @@
 import React from 'react'
-import { LocaleProvider } from 'antd'
-import { IntlProvider, addLocaleData } from 'react-intl'
+import { ConfigProvider } from 'antd'
+import { IntlProvider } from 'react-intl'
 import { connect } from 'react-redux'
 import portuguese from 'locales/pt-BR'
 import english from 'locales/en-US'
 import french from 'locales/fr-FR'
-import russian from 'locales/ru-RU'
-import chinese from 'locales/zh-CN'
-
-addLocaleData(portuguese.localeData)
-addLocaleData(english.localeData)
-addLocaleData(french.localeData)
-addLocaleData(russian.localeData)
-addLocaleData(chinese.localeData)
 
 const locales = {
   'pt-BR': portuguese,
   'en-US': english,
   'fr-FR': french,
-  'ru-RU': russian,
-  'zh-CN': chinese,
 }
 
 @connect(({ settings }) => ({ settings }))
@@ -31,11 +21,11 @@ class Localization extends React.Component {
     } = this.props
     const currentLocale = locales[locale]
     return (
-      <LocaleProvider locale={currentLocale.antdData}>
+      <ConfigProvider locale={currentLocale.antdData}>
         <IntlProvider locale={currentLocale.locale} messages={currentLocale.messages}>
           {children}
         </IntlProvider>
-      </LocaleProvider>
+      </ConfigProvider>
     )
   }
 }
