@@ -1,57 +1,67 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
+import React, { useCallback } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { useHistory } from 'react-router-dom'
 import { Form } from 'antd'
 
-import actions from 'redux/step/actions'
 import FormStepButtonsActions from 'components/Step/FormStepButtonsActions'
 import SkeletonForm from 'components/SkeletonForm/SkeletonForm'
+import passengerActions from 'redux/passengerDetail/actions'
 
-@Form.create()
-class PassengerForm extends Component {
-  constructor() {
-    super()
+const PassengerForm = ({ form, formSteps }) => {
+  const dispatch = useDispatch()
+  const history = useHistory()
+  const { current: currentStep } = useSelector(state => state.step)
+  const { isLoading } = useSelector(state => state.excursionDetail)
 
-    this.dispatchStep = this.dispatchStep.bind(this)
+  const onSaveFormAndAddNew = useCallback(() => {
+    saveAndRedirectTo(`${history.location.pathname}`)
+  }, [saveAndRedirectTo, history])
 
-    this.state = {
-      isLoading: true,
-    }
-  }
+  const onSubmit = useCallback(event => {
+    event.preventDefault()
 
-  componentDidMount() {
-    setTimeout(() => {
-      this.setState({ isLoading: false })
-    }, 1500)
-  }
+    saveAndRedirectTo(`${history.location.pathname}/list`)
+  }, [saveAndRedirectTo, history])
 
-  dispatchStep(step) {
-    const { dispatch } = this.props
-    dispatch({
-      type: actions.SET_STATE,
-      payload: { step },
+  const saveStepHandler = useCallback((fields, doSuccess) => {
+    form.validateFields(fields, { first: true }, (error, values) => {
+      if (!error) {
+        dispatch({ type: passengerActions.SET_PAYLOAD, payload: values })
+
+        doSuccess()
+      }
     })
-  }
+  }, [form, dispatch])
 
-  render() {
-    const { current, formSteps } = this.props
-    const { isLoading } = this.state
+  const saveAndRedirectTo = useCallback((redirect) => {
+    form.validateFields(async (error, values) => {
+      if (!error) {
+        const { keys, ...data } = values
+        await dispatch({ type: passengerActions.SAVE_PASSENGER, payload: data })
+        history.push(redirect)
+      }
+    })
+  }, [form, history, dispatch])
 
-    return (
-      <SkeletonForm isLoading={isLoading} rows={3}>
-        <Form layout="vertical" className="passenger-form">
-          {formSteps.map((x, i) => current === i && <x.component key={x.title} {...this.props} />)}
-
-          <div className="form-actions">
-            <FormStepButtonsActions lastStep={formSteps.length - 1} />
+  return (
+    <SkeletonForm isLoading={isLoading} rows={3}>
+      <Form id="passenger-form" hideRequiredMark colon={false} onSubmit={onSubmit}>
+        {formSteps.map((x, i) => (
+          <div key={x.title} style={{ display: currentStep === i ? 'block' : 'none' }}>
+            <x.component form={form} />
+            <div className="form-actions">
+              <FormStepButtonsActions
+                lastStep={formSteps.length - 1}
+                validationFields={x.fields}
+                onSaveStep={saveStepHandler}
+                onSaveFormAndAddNew={onSaveFormAndAddNew}
+              />
+            </div>
           </div>
-        </Form>
-      </SkeletonForm>
-    )
-  }
+        ))}
+      </Form>
+    </SkeletonForm>
+  )
 }
 
-const mapStateToProps = store => ({
-  current: store.step.current,
-})
-
-export default connect(mapStateToProps)(PassengerForm)
+export default Form.create()(PassengerForm)
