@@ -1,69 +1,104 @@
-import React, { Component } from 'react'
+import React, { useCallback } from 'react'
 import { Row, Col, Form, Select } from 'antd'
 
-import { busSeats, busStop } from 'mock/excursionSelects'
+import { useSelector } from 'react-redux'
 
-class PassengerPlace extends Component {
-  render() {
-    const { form } = this.props
+const PassengerPlace = (props) => {
+  const { form } = props
 
-    return (
-      <div>
-        <Row className="mb-5">
-          <Col xs={24} md={12}>
-            <b>Passageiro: </b>
-            <span>Fulano da Silva</span>
-          </Col>
-          <Col xs={24} md={12}>
-            <b>Tipo de passagem: </b>
-            <span>Normal</span>
-          </Col>
-        </Row>
+  const { payload: excursion } = useSelector((state) => state.excursionDetail)
 
-        <Row>
-          <Col xs={24} sm={8} lg={6}>
-            {/* TODO: transport ID */}
-            <Form.Item label="Tranporte">
-              {form.getFieldDecorator('transportId', { rules: [{ required: false }] })(
-                <Select size="default">
-                  <Select.Option key={1} value={1}>
-                    {/* TODO: */} TODO:
+  const vacancies =
+    excursion && excursion.transports
+      ? getVacancies(excursion.passengers, excursion.transports[0])
+      : []
+
+  const getStopPoints = useCallback(() => {
+    const { stopPoints = [] } = excursion
+    return stopPoints.map((x) => ({
+      value: x.id,
+      label: x.stopPoint,
+    }))
+  }, [excursion])
+
+  const getTransports = useCallback(() => {
+    const { transports = [] } = excursion
+    // TODO: criar enum para type de transport se não tiver, e adicionar no label
+    return transports.map((x) => ({
+      value: x.id,
+      label: `${x.type} ${x.plate} (${x.capacity})`,
+    }))
+  }, [excursion])
+
+  return (
+    <div>
+      {/* TODO: componentizar este header */}
+      <Row className="mb-5">
+        <Col xs={24} md={12}>
+          <b>Passageiro: </b>
+          <span>Fulano da Silva</span>
+        </Col>
+        <Col xs={24} md={12}>
+          <b>Tipo de passagem: </b>
+          <span>Normal</span>
+        </Col>
+      </Row>
+
+      <Row>
+        <Col xs={24} sm={8} lg={6}>
+          <Form.Item label="Transporte">
+            {form.getFieldDecorator('transportId', { rules: [{ required: false }] })(
+              <Select>
+                {getTransports().map((x) => (
+                  <Select.Option key={x.value} value={x.value}>
+                    {x.label}
                   </Select.Option>
-                </Select>,
-              )}
-            </Form.Item>
-          </Col>
-          <Col xs={24} sm={8} lg={6}>
-            <Form.Item label="Escolha do assento">
-              {form.getFieldDecorator('spot', { rules: [{ required: false }] })(
-                <Select size="default">
-                  {busSeats.map(x => (
-                    <Select.Option key={x.number} value={x.number} disabled={!x.free}>
-                      Poltrona {x.number} {x.free ? '' : ' - Reservado'}
-                    </Select.Option>
-                  ))}
-                </Select>,
-              )}
-            </Form.Item>
-          </Col>
+                ))}
+              </Select>,
+            )}
+          </Form.Item>
+        </Col>
+        <Col xs={24} sm={8} lg={6}>
+          <Form.Item label="Escolha do assento">
+            {form.getFieldDecorator('spot', { rules: [{ required: false }] })(
+              <Select>
+                {vacancies.map((x) => (
+                  <Select.Option key={x.number} value={x.number} disabled={!x.free}>
+                    Poltrona {x.number} {x.free ? '' : ' - Reservado'}
+                  </Select.Option>
+                ))}
+              </Select>,
+            )}
+          </Form.Item>
+        </Col>
 
-          <Col xs={24} sm={16} lg={18}>
-            <Form.Item label="Ponto de embarque">
-              {form.getFieldDecorator('stopPointId', { rules: [{ required: false }] })(
-                <Select size="default">
-                  {busStop.map(x => (
-                    <Select.Option key={x.id} value={x.id}>
-                      {x.address} - {x.time}
-                    </Select.Option>
-                  ))}
-                </Select>,
-              )}
-            </Form.Item>
-          </Col>
-        </Row>
-      </div>
-    )
-  }
+        <Col xs={24} sm={16} lg={12}>
+          <Form.Item label="Ponto de embarque">
+            {form.getFieldDecorator('stopPointId', { rules: [{ required: false }] })(
+              <Select>
+                {getStopPoints().map((x) => (
+                  <Select.Option key={x.value} value={x.value}>
+                    {x.label}
+                  </Select.Option>
+                ))}
+              </Select>,
+            )}
+          </Form.Item>
+        </Col>
+      </Row>
+    </div>
+  )
+}
+
+const getVacancies = (passengers, transports) => {
+  return Array(transports.capacity)
+    .fill(null)
+    .map((_, i) => {
+      const number = i + 1
+      const free = !passengers.some((p) => p.spot === number)
+
+      return { number, free }
+    })
 }
 
 export default PassengerPlace
