@@ -2,7 +2,7 @@ import actions from './actions'
 
 const initialState = {
   isLoading: false,
-  error: {},
+  error: false,
   payload: {},
 }
 
@@ -14,15 +14,14 @@ export default function reducer(state = initialState, action) {
     case actions.SAVE_EXCURSION:
       return { ...state, isLoading: true, payload: action.payload }
 
-    case actions.SAVE_EXCURSION_FAILURE:
-      state.isLoading = false
-      // TODO:
-      return state
+    case actions.GET_EXCURSION_BY_ID:
+      return { ...state, isLoading: true, payload: {} }
 
-    case actions.SAVE_EXCURSION_SUCCESS:
-      state.isLoading = false
-      // TODO:
-      return state
+    case actions.GET_EXCURSION_BY_ID_SUCCESS:
+      return { ...state, isLoading: false, payload: action.payload }
+
+    case actions.GET_EXCURSION_BY_ID_FAILURE:
+      return { ...state, isLoading: false, error: true }
 
     default:
       return state
