@@ -1,7 +1,14 @@
-import { all, takeEvery, call, put } from 'redux-saga/effects'
+import { all, takeEvery, call, put, takeLatest } from 'redux-saga/effects'
 import { notification } from 'antd'
 
-import actions, { saveExcursion, saveExcursionSuccess, saveExcursionFailure } from './actions'
+import actions, {
+  saveExcursion,
+  saveExcursionSuccess,
+  saveExcursionFailure,
+  getExcursionById,
+  getExcursionByIdSuccess,
+  getExcursionByIdFailure,
+} from './actions'
 
 export function* save({ payload }) {
   const result = yield call(saveExcursion, payload)
@@ -22,6 +29,25 @@ export function* save({ payload }) {
   }
 }
 
+export function* getById({ payload: id }) {
+  const getExcursion = getExcursionById(id)
+  const result = yield call(getExcursion.request)
+
+  if (result.response && result.response.data) {
+    yield put(getExcursionByIdSuccess(result.response.data))
+  } else {
+    notification.error({
+      message: 'Error',
+      description: 'Houve algum problema ao obter os dados da excursão!',
+    })
+    // const validationError = result && result.networkError && result.networkError.result.errors[0]
+    yield put(getExcursionByIdFailure())
+  }
+}
+
 export default function* rootSaga() {
-  yield all([takeEvery(actions.SAVE_EXCURSION, save)])
+  yield all([
+    takeEvery(actions.SAVE_EXCURSION, save),
+    takeLatest(actions.GET_EXCURSION_BY_ID, getById),
+  ])
 }
